requirements.txt
numpy
opencv-python
pandas
python-dotenv
dotenv
scipy
shapely
mediapipe
pytorch